Showing my math
Are there any mountains left to climb in this great wasteland we call Verena? Are there any interesting questions left to ask? Is it possible we have already hit the ceiling on human understanding? Does tidyverse make it too easy to do analytical work? Where is my Toki Underground delivery?